Round 9 (#9)
* refactor: 결제 시 포인트 이중 검증 수정

* refactor: 캐시 및 이벤트 스트림 아키텍처 개선

* refactor: 좋아요 이벤트 처리 로직 개선

* feat: 상품 랭킹 시스템 구현

* test: 기존 테스트 버그 수정 및 랭킹 테스트 추가

* refactor: 랭킹 API의 date 파라미터 required=true로 변경

* refactor: 랭킹 NPE(NullPointerException) 방어 로직 추가

* refactor: 랭킹 스코어 계산 부정합 코드 수정

* refactor: commerce-streamer 패키지 구조 변경

+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/event/outbox/OutboxRelay.java
@@ -102,9 +102,9 @@ public class OutboxRelay {
 				.eventType(outbox.getEventType())
 				.aggregateId(outbox.getAggregateId())
 				.timestamp(LocalDateTime.now())
-				.version(outbox.getEventVersion() != null
-					? outbox.getEventVersion().intValue()
-					: 1)
+                .version(outbox.getEventVersion() != null
+                    ? outbox.getEventVersion()
+                    : 1L)
 				.payload(payload)
 				.build();
